+++ ios/TargetAppRN.xcodeproj/xcshareddata/xcschemes/TargetAppRN-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBase-tvOS.app"
-               BlueprintName = "ReactNativeBase-tvOS"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BuildableName = "TargetAppRN-tvOS.app"
+               BlueprintName = "TargetAppRN-tvOS"
+               ReferencedContainer = "container:TargetAppRN.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBase-tvOSTests.xctest"
-               BlueprintName = "ReactNativeBase-tvOSTests"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BuildableName = "TargetAppRN-tvOSTests.xctest"
+               BlueprintName = "TargetAppRN-tvOSTests"
+               ReferencedContainer = "container:TargetAppRN.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBase-tvOSTests.xctest"
-               BlueprintName = "ReactNativeBase-tvOSTests"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BuildableName = "TargetAppRN-tvOSTests.xctest"
+               BlueprintName = "TargetAppRN-tvOSTests"
+               ReferencedContainer = "container:TargetAppRN.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBase-tvOS.app"
-            BlueprintName = "ReactNativeBase-tvOS"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BuildableName = "TargetAppRN-tvOS.app"
+            BlueprintName = "TargetAppRN-tvOS"
+            ReferencedContainer = "container:TargetAppRN.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBase-tvOS.app"
-            BlueprintName = "ReactNativeBase-tvOS"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BuildableName = "TargetAppRN-tvOS.app"
+            BlueprintName = "TargetAppRN-tvOS"
+            ReferencedContainer = "container:TargetAppRN.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBase-tvOS.app"
-            BlueprintName = "ReactNativeBase-tvOS"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BuildableName = "TargetAppRN-tvOS.app"
+            BlueprintName = "TargetAppRN-tvOS"
+            ReferencedContainer = "container:TargetAppRN.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
