+++ ios/TargetAppRN.xcodeproj/xcshareddata/xcschemes/TargetAppRN.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "ReactNativeBase.app"
-               BlueprintName = "ReactNativeBase"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BuildableName = "TargetAppRN.app"
+               BlueprintName = "TargetAppRN"
+               ReferencedContainer = "container:TargetAppRN.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ReactNativeBaseTests.xctest"
-               BlueprintName = "ReactNativeBaseTests"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BuildableName = "TargetAppRNTests.xctest"
+               BlueprintName = "TargetAppRNTests"
+               ReferencedContainer = "container:TargetAppRN.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ReactNativeBaseTests.xctest"
-               BlueprintName = "ReactNativeBaseTests"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BuildableName = "TargetAppRNTests.xctest"
+               BlueprintName = "TargetAppRNTests"
+               ReferencedContainer = "container:TargetAppRN.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeBase.app"
-            BlueprintName = "ReactNativeBase"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BuildableName = "TargetAppRN.app"
+            BlueprintName = "TargetAppRN"
+            ReferencedContainer = "container:TargetAppRN.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeBase.app"
-            BlueprintName = "ReactNativeBase"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BuildableName = "TargetAppRN.app"
+            BlueprintName = "TargetAppRN"
+            ReferencedContainer = "container:TargetAppRN.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeBase.app"
-            BlueprintName = "ReactNativeBase"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BuildableName = "TargetAppRN.app"
+            BlueprintName = "TargetAppRN"
+            ReferencedContainer = "container:TargetAppRN.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
