+++ ios/TargetAppRN/Images.xcassets/AppIcon.appiconset/Contents.json
