+++ ios/TargetAppRN/Images.xcassets/Contents.json
